% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage_ratio.R
\name{coverage_ratio}
\alias{coverage_ratio}
\title{Coverage ratio between training and test sets.}
\usage{
coverage_ratio(X_train, X_test)
}
\arguments{
\item{X_train}{The predictors in the training set, an n-by-p matrix.}

\item{X_test}{The predictors in the training set, a matrix with the same number of columns as X.}
}
\value{
The Coverage ratio between train and test sets.
}
\description{
Coverage Ratio is a metric that measures the similarity between the training set and the test set.
    Based on the dimensional reduction graph of tsne, we calculated how many data points in the test set have k-nearest neighbors that contain the training set, that is,
    they are covered by the training set. If CR < 0.5, we believe that this test set is significantly different from the training set and may not be applicable to the corresponding model.
}
\examples{
    X_train <- matrix(rnorm(2000), 200, 10)
    X_test <- matrix(rnorm(1000, 1), 100, 10)
    coverage_ratio(X_train, X_test)
}
