% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_improvement.R
\name{ps_probs}
\alias{ps_probs}
\title{Structure improvement evaluation based on predicted probability.}
\usage{
ps_probs(X, Y, predicted_probs, cores = 1, BootNumber = 1000)
}
\arguments{
\item{X}{The predictors in the training set, which is an n times p dimensional matrix.}

\item{Y}{The true response in the training set, which is an n-dimensional vector.}

\item{cores}{The number of threads used for parallel computing. It can only be 1 in Windows environment}

\item{BootNumber}{The number of bootstrap steps in the calculation of p_s.}

\item{Y2}{The predicted probability of the trained model on the training set X.}
}
\value{
The structure improvement index p_s. If p_s < 0.05, the model fit is not good enough and we need to increase the complexity of the model.
}
\description{
Compute the structure improvement index p_s based on predicted probability. When Y is a binary variable,
    we use the probability vector given by the trained model to calculate the model structure improvement.
}
\examples{
X <- matrix(rnorm(2000), 100, 20)
    tmp <- X \%*\% rnorm(20) + rnorm(20)
    probabilities <- 1 / (1 + exp(-tmp))
    Y <- rbinom(100, 1, probabilities)
    model <- glm(Y ~ X - 1, family = binomial)
    predicted_probabilities <- predict(model, newdata = data.frame(X), type = "response")
    ps_probs(X, Y, predicted_probabilities, 1, 100)
}
